Fix inv_vref_scan_lund bug (#394)
* fixed bug causing fit error in inv_vref_scan_lund

* fix so that slopes exactly at the flat threshold are correctly excluded from the median calculation

+++ app/tool/algorithm/inv_vref_lund.h
@@ -36,9 +36,9 @@ class DataFitter {
   };
   std::vector<Point> linear_;
   std::vector<Point> nonlinear_;
-  int LH_median_;
+  double LH_median_;
   double LH_std_median_;
-  int RH_median_;
+  double RH_median_;
 };
 
 /**
